Standardized naming.
All files renamed to kern/alg/kern (instead of kern/alg/whatever).
Fixed Makefiles, source, and aux files accordingly.
