+++ backprop/backprop/Makefile
@@ -1,11 +1,11 @@
-net: net.c net.h ../../common/harness.c
-	$(CC) $(CFLAGS) -o net net.c ../../common/harness.c -lm
+backprop: backprop.c backprop.h ../../common/harness.c
+	$(CC) $(CFLAGS) -o backprop backprop.c ../../common/harness.c -lm
 
-run: net input.data check.data
-	./net input.data check.data
+run: backprop input.data check.data
+	./backprop input.data check.data
 
-hls: net.c net.h
+hls: backprop.c backprop.h
 	vivado_hls hls.tcl
 
 clean:
-	rm -f net
+	rm -f backprop
