+++ backprop/backprop/hls.tcl
@@ -1,6 +1,6 @@
-open_project net_hls
+open_project backprop_hls
 
-add_files net.c
+add_files backprop.c
 add_files input.data
 add_files check.data
 add_files -tb ../../common/harness.c
@@ -11,7 +11,7 @@ open_solution -reset solution
 set_part virtex7
 create_clock -period 10
 
-#source ./net_dir
+#source ./backprop_dir
 config_rtl -reset all -reset_level low
 
 csynth_design
