+++ bfs/bulk/bfs_dir
