+++ bfs/bulk/generate.c
@@ -23,7 +23,7 @@
 #define D 25
 #endif
 
-#include "bulk.h"
+#include "bfs.h"
 // Fake benchmark function to satisfy the extern
 
 void bfs(node_t nodes[N_NODES], edge_t edges[N_EDGES], node_index_t starting_node, level_t level[N_NODES], edge_index_t level_counts[N_LEVELS]) { }
