+++ bfs/bulk/hls.tcl
@@ -1,6 +1,6 @@
-open_project bulk_syn
+open_project bfs_syn
 
-add_files bulk.c
+add_files bfs.c
 add_files input.data
 add_files check.data
 add_files -tb ../../common/harness.c
@@ -10,7 +10,7 @@ set_top bfs
 open_solution -reset solution
 set_part virtex7
 create_clock -period 10
-#source ./bulk_dir
+#source ./bfs_dir
 #config_rtl -reset all -reset_level low
 csynth_design
 cosim_design -rtl verilog -tool modelsim -trace_level all
