+++ sort/merge/Makefile
@@ -1,11 +1,11 @@
-merge: merge.c merge.h ../../common/harness.c
-	$(CC) $(CFLAGS) -o merge merge.c ../../common/harness.c
+sort: sort.c sort.h ../../common/harness.c
+	$(CC) $(CFLAGS) -o sort sort.c ../../common/harness.c
 
-run: merge input.data check.data
-	./merge input.data check.data
+run: sort input.data check.data
+	./sort input.data check.data
 
-hls: merge.c merge.h
+hls: sort.c sort.h
 	vivado_hls hls.tcl
 
 clean:
-	rm -f merge
+	rm -f sort
