+++ spmv/ellpack/hls.tcl
@@ -1,6 +1,6 @@
-open_project ellpack_syn
+open_project spmv_syn
 
-add_files ellpack.c
+add_files spmv.c
 add_files input.data
 add_files check.data
 add_files -tb ../../common/harness.c
@@ -10,7 +10,7 @@ set_top ellpack
 open_solution -reset solution
 set_part virtex7
 create_clock -period 10
-source ./ellpack_dir
+source ./spmv_dir
 csynth_design
 cosim_design -rtl verilog -tool modelsim -trace_level all
