+++ stencil/stencil3d/hls.tcl
@@ -1,6 +1,6 @@
-open_project stencil3d_syn
+open_project stencil_syn
 
-add_files stencil3d.c
+add_files stencil.c
 add_files input.data
 add_files check.data
 add_files -tb ../../common/harness.c 
@@ -10,7 +10,7 @@ open_solution -reset solution
 
 set_part virtex7
 create_clock -period 10
-source ./stencil3d_dir
+source ./stencil_dir
 
 csynth_design
 cosim_design -rtl verilog -tool modelsim -trace_level all
