+++ stencil/stencil3d/stencil.c
@@ -27,31 +27,35 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-Implementation based on algorithm described in:
-The cache performance and optimizations of blocked algorithms
-M. D. Lam, E. E. Rothberg, and M. E. Wolf
-ASPLOS 1991
+implementation based on algorithm described in:
+"Stencil computation optimization and auto-tuning on state-of-the-art multicore architectures"
+K. Datta, M. Murphy, V. Volkov, S. Williams, J. Carter, L. Oliker, D. Patterson, J. Shalf, K. Yelick
+SC 2008
 */
 
-#include "bbgemm.h"
+#include "stencil.h"
 
-void bbgemm(TYPE m1[N], TYPE m2[N], TYPE prod[N]){
-    int i, k, j, jj, kk, temp_x;
-    int i_row, k_row;
-    TYPE mul;
+void stencil3d(TYPE C0, TYPE C1, TYPE orig[size], TYPE sol[size]) {
+    int i, j, k;
+    TYPE sum1, sum2, mul1, mul2, diff;
 
-    loopjj:for (jj = 0; jj < row_size; jj += block_size){
-        loopkk:for (kk = 0; kk < row_size; kk += block_size){
-            loopi:for ( i = 0; i < row_size; ++i){
-                loopk:for (k = 0; k < block_size; ++k){
-                    i_row = i * row_size;
-                    k_row = (k  + kk) * row_size;
-                    temp_x = m1[i_row + k + kk];
-                    loopj:for (j = 0; j < block_size; ++j){
-                        mul = temp_x * m2[k_row + j + jj];
-                        prod[i_row + j + jj] += mul;
-                    }
-                }
+    loop_height : for(i = 1; i < height_size - 1; i++){
+        loop_col : for(j = 1; j < col_size - 1; j++){
+            loop_row : for(k = 1; k < row_size - 1; k++){
+                sum1 = orig[indx(row_size, col_size, i, j, k + 1)] +
+                       orig[indx (row_size, col_size, i, j, k - 1)] +
+                       orig[indx (row_size, col_size, i, j + 1, k)] +
+                       orig[indx (row_size, col_size, i, j - 1, k)] +
+                       orig[indx (row_size, col_size, i + 1, j, k)] +
+                       orig[indx (row_size, col_size, i - 1, j, k)];
+
+                sum2 = orig[indx (row_size, col_size, i, j, k)];
+
+                mul2 = sum2 * C0;
+                mul1 = sum1 * C1;
+                diff = mul1 - mul2;
+
+                sol[indx(row_size, col_size, i, j, k)] = diff;
             }
         }
     }
