+++ stencil/stencil3d/stencil.h
