+++ stencil/stencil3d/stencil_dir
